notebooks/framework-tests/insurance-semantic-kernel/pyproject.toml
[project]
name = "agentic-functions"
version = "0.1.0"
description = "description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "azure-ai-evaluation>=1.0.0b5",
    "azure-cosmos>=4.7.0",
    "azure-identity>=1.19.0",
    "azure-search-documents>=11.5.2",
    "azure-search>=1.0.0b2",
    "ipykernel>=6.29.5",
    "nest-asyncio>=1.6.0",
    "openai>=1.52.2",
    "pandas>=2.2.3",
    "pydantic-core>=2.23.4",
    "pydantic-settings>=2.6.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "semantic-kernel>=1.13.0",
]
